refactor: use useBorrowTransaction hook
